Feature: Novels of 2012 mini series selectable in Romantauschbörse (#469)
* Add new book type to allowed list

Included BookType::ZweiTausendZwölfDasJahrDerApokalypse in the allowed book types array to support this new type in RomantauschController.

* Add test data for 2012 Roman book type

Added a Book instance with type ZweiTausendZwölfDasJahrDerApokalypse to the putBookData() method in RomantauschControllerTest.php to support tests involving this book type.

* Add test for 2012 mini series book offer creation

Introduces a new test to verify that creating an offer for the '2012 Mini Series' correctly inserts the expected entry into the 'book_offers' table.

* Add test for 2012 mini series book request creation

Introduces a new test to verify that submitting a book request for the '2012 Mini Series' correctly creates an entry in the database with the expected attributes.